添加重跑实例记录时间问题

+++ spark-yun-backend/spark-yun-modules/src/main/java/com/isxcode/spark/modules/work/run/impl/ExecuteSqlExecutor.java
@@ -1,5 +1,6 @@
 package com.isxcode.spark.modules.work.run.impl;
 
+import com.alibaba.fastjson.JSON;
 import com.isxcode.spark.api.datasource.dto.ConnectInfo;
 import com.isxcode.spark.api.work.constants.WorkLog;
 import com.isxcode.spark.api.work.constants.WorkType;
@@ -11,9 +12,11 @@ import com.isxcode.spark.modules.datasource.mapper.DatasourceMapper;
 import com.isxcode.spark.modules.datasource.repository.DatasourceRepository;
 import com.isxcode.spark.modules.datasource.source.DataSourceFactory;
 import com.isxcode.spark.modules.datasource.source.Datasource;
+import com.isxcode.spark.modules.secret.repository.SecretKeyRepository;
 import com.isxcode.spark.modules.work.entity.WorkInstanceEntity;
 import com.isxcode.spark.modules.work.repository.WorkInstanceRepository;
 import com.isxcode.spark.modules.work.run.WorkExecutor;
+import com.isxcode.spark.modules.work.run.WorkInfo;
 import com.isxcode.spark.modules.work.run.WorkRunContext;
 import com.isxcode.spark.modules.work.sql.SqlCommentService;
 import com.isxcode.spark.modules.work.sql.SqlFunctionService;
@@ -47,10 +50,13 @@ public class ExecuteSqlExecutor extends WorkExecutor {
 
     private final DatasourceMapper datasourceMapper;
 
+    private final SecretKeyRepository secretKeyRepository;
+
     public ExecuteSqlExecutor(WorkInstanceRepository workInstanceRepository, DatasourceRepository datasourceRepository,
         WorkflowInstanceRepository workflowInstanceRepository, SqlCommentService sqlCommentService,
         SqlValueService sqlValueService, SqlFunctionService sqlFunctionService, AlarmService alarmService,
-        DataSourceFactory dataSourceFactory, DatasourceMapper datasourceMapper) {
+        DataSourceFactory dataSourceFactory, DatasourceMapper datasourceMapper,
+        SecretKeyRepository secretKeyRepository) {
 
         super(workInstanceRepository, workflowInstanceRepository, alarmService, sqlFunctionService);
         this.datasourceRepository = datasourceRepository;
@@ -59,6 +65,7 @@ public class ExecuteSqlExecutor extends WorkExecutor {
         this.sqlFunctionService = sqlFunctionService;
         this.dataSourceFactory = dataSourceFactory;
         this.datasourceMapper = datasourceMapper;
+        this.secretKeyRepository = secretKeyRepository;
     }
 
     @Override
@@ -111,17 +118,31 @@ public class ExecuteSqlExecutor extends WorkExecutor {
 
             statement.setQueryTimeout(1800);
 
-            // 去掉sql中的注释
-            String sqlNoComment = sqlCommentService.removeSqlComment(workRunContext.getScript());
+            // 判断实例中是否有脚本
+            String script;
+            if (workInstance.getWorkInfo() == null) {
+                // 去掉sql中的注释
+                String sqlNoComment = sqlCommentService.removeSqlComment(workRunContext.getScript());
 
-            // 解析上游参数
-            String jsonPathSql = parseJsonPath(sqlNoComment, workInstance);
+                // 解析上游参数
+                String jsonPathSql = parseJsonPath(sqlNoComment, workInstance);
 
-            // 翻译sql中的系统变量
-            String parseValueSql = sqlValueService.parseSqlValue(jsonPathSql);
+                // 翻译sql中的系统变量
+                String parseValueSql = sqlValueService.parseSqlValue(jsonPathSql);
 
-            // 翻译sql中的系统函数
-            String script = sqlFunctionService.parseSqlFunction(parseValueSql);
+                // 翻译sql中的系统函数
+                try {
+                    script = sqlFunctionService.parseSqlFunction(parseValueSql);
+                } catch (Exception e) {
+                    throw new WorkRunException(
+                        LocalDateTime.now() + WorkLog.ERROR_INFO + "系统函数异常\n" + e.getMessage() + "\n");
+                }
+
+                workInstance.setWorkInfo(JSON.toJSONString(WorkInfo.builder().script(script).build()));
+            } else {
+                WorkInfo workInfo = JSON.parseObject(workInstance.getWorkInfo(), WorkInfo.class);
+                script = workInfo.getScript();
+            }
 
             // 清除脚本中的脏数据
             List<String> sqls =
